// apps/templates/wordsmith/types/generated/types.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  DateTime: { input: any; output: any; }
  JSON: { input: any; output: any; }
  JSONObject: { input: any; output: any; }
};

/** Asset */
export type Asset = {
  __typename?: 'Asset';
  blurhash: Scalars['String']['output'];
  cloudfareId?: Maybe<Scalars['String']['output']>;
  height?: Maybe<Scalars['Float']['output']>;
  id: Scalars['ID']['output'];
  /** This is the key from the bucket */
  key: Scalars['String']['output'];
  lastReferenceCheckedAt?: Maybe<Scalars['String']['output']>;
  mimeType: Scalars['String']['output'];
  originalFileName: Scalars['String']['output'];
  owner: User;
  size: Scalars['Float']['output'];
  variants: Array<Scalars['String']['output']>;
  width?: Maybe<Scalars['Float']['output']>;
};
// ...
export type Page = {
  __typename?: 'Page';
  bannerImage?: Maybe<Asset>;
  body?: Maybe<Scalars['String']['output']>;
  children: Array<Page>;
  createdAt: Scalars['DateTime']['output'];
  draftBody?: Maybe<Scalars['String']['output']>;
  id: Scalars['ID']['output'];
  isFeatured: Scalars['Boolean']['output'];
  isPublic?: Maybe<Scalars['Boolean']['output']>;
  jsonBody?: Maybe<Scalars['JSON']['output']>;
  jsonDraftBody?: Maybe<Scalars['JSON']['output']>;
  lastPublishedAt?: Maybe<Scalars['DateTime']['output']>;
  metaDescription?: Maybe<Scalars['String']['output']>;
  metaTitle?: Maybe<Scalars['String']['output']>;
  name: Scalars['String']['output'];
  owner: Scalars['ID']['output'];
  parent?: Maybe<Page>;
  priority: Scalars['Int']['output'];
  site: Scalars['ID']['output'];
  slug: Scalars['String']['output'];
  updatedAt: Scalars['DateTime']['output'];
};

export enum PageFilterEnum {
  Any = 'ANY',
  Private = 'PRIVATE',
  Public = 'PUBLIC'
}

export type PageResponse = {
  __typename?: 'PageResponse';
  data?: Maybe<Page>;
  message?: Maybe<Scalars['String']['output']>;
  statusCode?: Maybe<Scalars['Int']['output']>;
  success?: Maybe<Scalars['Boolean']['output']>;
};